# api/openapi.yaml
openapi: 3.1.0
info:
  title: AURA — Ashesi University Resource Allocation API
  version: "1.0.0"
  description: >
    Contract-first REST API for AURA, Ashesi University's resource allocation and
    reservation platform (§8). JSON over HTTPS,
    versioned under /api/v1. Errors use application/problem+json (RFC 9457) with a
    stable machine `code`. Cursor pagination via ?limit & ?cursor. This document is
    the source of truth from which the web and mobile typed clients are generated.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.aura.ashesi.edu
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/bookings:
    post:
      tags: [Bookings]
      summary: Submit a booking request (FR7)
      parameters:
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [room_id, purpose, attendee_count, starts_at, ends_at]
              properties:
                room_id: { type: string, format: uuid }
                purpose: { type: string }
                attendee_count: { type: integer, minimum: 1 }
                starts_at: { type: string, format: date-time }
                ends_at: { type: string, format: date-time }
      responses:
        "201":
          description: Created (PENDING)
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking: { $ref: "#/components/schemas/Booking" }
                  competing_pending: { type: integer }
        "422": { $ref: "#/components/responses/Problem", description: "Conflicts with lecture/maintenance, capacity, or past" }
    get:
      tags: [Bookings]
      summary: List bookings (scope mine|pending|all)
      description: >
        Returns a cursor-paginated {data, next_cursor} envelope. For scope=pending
        (the approvals queue), each data item is a BookingApprovability object —
        the booking enriched with nested room/requester plus the §11/FR8 blockers
        explaining whether it can be approved. For scope=mine|all, each data item
        is a plain Booking.
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, pending, all] } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/BookingStatus" } }
        - { name: room_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: >
            OK. data is an array of Booking (scope mine|all) or
            BookingApprovability (scope pending).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      oneOf:
                        - { $ref: "#/components/schemas/Booking" }
                        - { $ref: "#/components/schemas/BookingApprovability" }
                  next_cursor: { type: string, nullable: true }
...
  /api/v1/reports/utilisation:
    get:
      tags: [Reports]
      summary: Room utilisation (FR12). format=csv for export.
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: building_id, in: query, schema: { type: string, format: uuid } }
        - { name: format, in: query, schema: { type: string, enum: [json, csv] } }
      responses: { "200": { description: OK } }
...
  /api/v1/reports/overview:
    get:
      tags: [Reports]
      summary: Admin overview dashboard aggregate (FR12). Defaults to the last 30 days.
      description: >
        Single aggregate report powering the admin overview dashboard. When from/to
        are omitted it defaults to the last 30 days (to = today, from = today − 30d).
        In-range counts are scoped by starts_at within [from, to); current-state KPIs
        (pending, active rooms/users, buildings) ignore the range.
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OverviewReport" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: __Host-access }
  parameters:
...
    Limit: { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Stable machine-readable error code }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
    Role: { type: string, enum: [SYSTEM_ADMIN, TIMETABLE_ADMIN, BOOKING_OFFICER, REQUESTER] }
    RoomType: { type: string, enum: [LECTURE_HALL, LAB, SEMINAR_ROOM, AUDITORIUM, CONFERENCE_ROOM] }
    BookingStatus: { type: string, enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXPIRED] }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        full_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        department_id: { type: string, format: uuid, nullable: true }
        status: { type: string }
        mfa_enabled: { type: boolean }
...
    Room:
      type: object
      properties:
        id: { type: string, format: uuid }
        room_code: { type: string }
        name: { type: string }
        building_id: { type: string, format: uuid }
        capacity: { type: integer }
        room_type: { $ref: "#/components/schemas/RoomType" }
        status: { type: string }
        building_code: { type: string }
        building_name: { type: string }
        image_url: { type: string, nullable: true }
        image_public_id: { type: string, nullable: true }
        gallery_urls: { type: array, items: { type: string } }
        gallery_public_ids: { type: array, items: { type: string } }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        room_id: { type: string, format: uuid }
        requested_by: { type: string, format: uuid }
        purpose: { type: string }
        attendee_count: { type: integer }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/BookingStatus" }
        review_note: { type: string, nullable: true }
        # Optional nested relations, returned on the enriched approvals projection.
        room: { allOf: [ { $ref: "#/components/schemas/Room" } ], nullable: true }
        requester: { allOf: [ { $ref: "#/components/schemas/User" } ], nullable: true }
    ApprovalBlocker:
      type: object
      description: A reason a pending booking can or cannot be approved (§11/FR8).
      required: [kind, message]
      properties:
        kind:
          type: string
          enum: [LECTURE, MAINTENANCE, APPROVED_BOOKING, COMPETING_PENDING, CAPACITY, IN_PAST]
        message: { type: string }
        starts_at: { type: string, format: date-time, nullable: true }
        ends_at: { type: string, format: date-time, nullable: true }
    BookingApprovability:
      type: object
      description: >
        A pending booking enriched with approvability info, used on the approvals
        queue. can_approve is true only when there are no hard blockers (IN_PAST,
        CAPACITY, LECTURE, MAINTENANCE, APPROVED_BOOKING); competing pendings are
        informational and never block.
      required: [booking, can_approve, blockers, competing_pending_count]
      properties:
        booking: { $ref: "#/components/schemas/Booking" }
        can_approve: { type: boolean }
        blockers:
          type: array
          items: { $ref: "#/components/schemas/ApprovalBlocker" }
        competing_pending_count: { type: integer }
...
    LabelCount:
      type: object
      required: [label, count]
      properties:
        label: { type: string }
        count: { type: integer }
    OverviewReport:
      type: object
      description: Aggregate report powering the admin overview dashboard (FR12).
      required: [range, kpis, status_breakdown, by_room_type, by_building, top_rooms, series, peak_hours]
      properties:
        range:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: date }
            to: { type: string, format: date }
        kpis:
          type: object
          required: [total_bookings, pending, approved, rejected, cancelled, expired, active_rooms, active_users, buildings, avg_utilisation_pct, approval_rate_pct]
          properties:
            total_bookings: { type: integer, description: Bookings with starts_at in range }
            pending: { type: integer, description: Current count of PENDING bookings (ignores range) }
            approved: { type: integer }
            rejected: { type: integer }
            cancelled: { type: integer }
            expired: { type: integer }
            active_rooms: { type: integer, description: Current rooms WHERE status='ACTIVE' }
            active_users: { type: integer, description: Current users WHERE status='ACTIVE' }
            buildings: { type: integer, description: Current building count }
            avg_utilisation_pct: { type: number, format: float }
            approval_rate_pct: { type: number, format: float, description: approved/(approved+rejected), 1 decimal }
        status_breakdown:
          type: array
          description: Bookings by status in range; all 5 statuses, zeros included.
          items: { $ref: "#/components/schemas/LabelCount" }
        by_room_type:
          type: array
          description: Bookings by room.room_type in range.
          items: { $ref: "#/components/schemas/LabelCount" }
        by_building:
          type: array
          description: Bookings by building.code in range.
          items: { $ref: "#/components/schemas/LabelCount" }
        top_rooms:
          type: array
          description: Top 8 rooms by utilisation_pct desc over the range.
          items:
            type: object
            required: [room_code, room_name, utilisation_pct, booked_hours]
            properties:
              room_code: { type: string }
              room_name: { type: string }
              utilisation_pct: { type: number, format: float }
              booked_hours: { type: number, format: float }
        series:
          type: array
          description: One point per calendar day in range (institution tz); empty days are zero.
          items:
            type: object
            required: [date, submitted, approved]
            properties:
              date: { type: string, format: date }
              submitted: { type: integer, description: Bookings created that day }
              approved: { type: integer, description: Bookings created that day whose status is APPROVED }
        peak_hours:
          type: array
          description: Bookings by local start hour-of-day 0..23 in range; all 24 hours included.
          items:
            type: object
            required: [hour, count]
            properties:
              hour: { type: integer, minimum: 0, maximum: 23 }
              count: { type: integer }